add namespace to build-standard-function-term
Without the explicit gamma.api namespace, when
build-standard-function-term is used from within an eval, the current
namespace will be gamma.api$macros and the function does not exist in
that namespace, so it fails.  At least, that's how it was explained to
me on twitter by Mike Fikes.  https://twitter.com/mfikes/status/758314635154059264

I don't have a good explanation for why this works in the default case, though.